ejercicio 13 corregido

+++ ejercicios01/src/ejercicio13/CalculadoraNotas.java
@@ -2,7 +2,7 @@ package ejercicio13;
 
 public class CalculadoraNotas {
 
-    public double segundaNota(double primeraNota, double notaDeseada) {
-        return (notaDeseada - primeraNota * 0.4) / 0.6;
+    public double segundaNota(float primeraNota, float notaDeseada) {
+        return (notaDeseada - (primeraNota * 0.4)) / 0.6;
     }
 }
